JWT token implemented in CRUD

+++ backend/controllers/goalController.js
@@ -1,12 +1,13 @@
 import asyncHandler from "express-async-handler"
 import Goal from "../model/goalModel.js"
+import User from "../model/userModel.js"
 
 
 // @desc get goals
 // @route GET /api/goals
 // @access private 
 export const getGoals = asyncHandler(async(req,res) => {
-    const goals = await Goal.find({})
+    const goals = await Goal.find({ user: req.user.id })
     res.status(200).json(goals)
 })
 
@@ -20,7 +21,8 @@ export const setGoals = asyncHandler(async(req,res) => {
     }
     
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
     res.status(200).json(goal)
 })
@@ -34,6 +36,19 @@ export const updateGoals = asyncHandler(async(req,res) => {
      if(!goal){
         res.status(400)
         throw new Error('Goal not found')    
+    }
+
+    const user = await User.findById(req.user.id)
+    // check for user
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // Make sure logged in user matches the goal user
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('User not authorized')
     }
 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -51,6 +66,19 @@ export const removeGoals = asyncHandler(async(req,res) => {
      if(!goal){
         res.status(400)
         throw new Error('Goal not found')    
+    }
+
+        const user = await User.findById(req.user.id)
+    // check for user
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // Make sure logged in user matches the goal user
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('User not authorized')
     }
 
     const deletedGoal = await Goal.findByIdAndDelete(req.params.id)
